refactor: renamed skybox file
