+++ Engine/src/graphics/resources/vk_skybox.cpp
@@ -1,4 +1,4 @@
-#include "graphics/resources/skybox.hpp"
+#include "graphics/resources/vk_skybox.hpp"
 
 #include "application.hpp"
