+++ Engine/src/graphics/resources/vk_skybox.hpp
